bluetooth: getDevices() implementation
This change implements getDevices() which returns a list of
WebBluetoothDevice objects that the current site has permission to
access. If the kWebBluetoothNewPermissionsBackend flag is enabled, the
list of devices will contain all of the permitted devices. If the flag
is not enabled, then the list of devices will contain the permitted
devices that are currently connected to the system.

Design doc:
https://docs.google.com/document/d/1h3uAVXJARHrNWaNACUPiQhLt7XI-fFFQoARSs1WgMDM/edit#heading=h.5ugemo7p04z9

Bug: 577953
Change-Id: I9785f24ee46ac634b6a96d6146f54da37d132a4e