Avoid scheduling invalidation sets from style recalc.
Sadly, we do SetInlineStyleProperty() for UA shadow DOM during style
recalc in some cases. Make sure that we don't schedule any invalidation
while in style reacalc as that would leave use with a dirty tree for
style invalidations after UpdateStyle().

This fixes a regression introduced in [1] where we unified the
invalidation code for the various ways of changing the inline style.

[1] https://crrev.com/2faf55836255e6762141712d943c47f3208c8ac4

Bug: 1057210
Change-Id: I0edab4acc03baefe0fb386c15dce78e0f8e041ff